requirements.txt
openai==1.30.1
flask==0.12
boto3==1.34.107
transformers==2.11.0
torch==1.10.0
pinecone-client=4.1.0
langchain=0.1.20
google-cloud=0.34.0
numpy==1.0.3
opencv-python==3.4.1.14
pillow=1.6
matplotlib==3.6.3
tensorflow-gpu==2.6.0
librosa=0.10.2
soundfile=0.12.1
pandas=2.2.2
scikit-learn
scipy==1.10.0